pyproject.toml
[project]
name = "triton-b64-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "requests>=2.32.4",
]
